\name{genotypes.poly}
\alias{genotypes.poly}
\docType{data}
\title{
Sample Barley Genotypes
}
\description{
Twenty diploid barley genotypes, with twenty markers each, for to demonstrate functions in the GeneticSubsetter package, formatted for CoreSetterPoly.
}
\usage{data(genotypes)}
\format{
   Rows are genotypes, and each set of two columns is one locus.
}
\source{
Triticeae Coordinated Agricultural Project (T-CAP) (http://triticeaetoolbox.org)
}
\examples{
data("genotypes")
str(genotypes.poly)
}
\keyword{datasets}
