\name{genotypes}
\alias{genotypes}
\docType{data}
\title{
Sample Barley Genotypes
}
\description{
Twenty barley genotypes, with twenty markers each, for to demonstrate functions in the GeneticSubsetter package.
}
\usage{data(genotypes)}
\format{
  A data frame where columns are gentoypes, and rows are markers.
}

\source{
Triticeae Coordinated Agricultural Project (T-CAP) (http://triticeaetoolbox.org)
}
\examples{
data("genotypes")
str(genotypes)
}
\keyword{datasets}
