\name{genotypes}
\alias{genotypes}
\docType{data}
\title{
Sample Barley Genotypes
}
\description{
Twenty diploid barley genotypes, with twenty markers each, for to demonstrate functions in the GeneticSubsetter package.
}
\usage{data(genotypes)}
\format{
  Columns are gentoypes, and rows are markers, formatted for the CoreSetter, CoreSetterCombined, CoreSetOptimizer, HET, MTK.
}
\source{
Triticeae Coordinated Agricultural Project (T-CAP) (http://triticeaetoolbox.org)
}
\examples{
data("genotypes")
str(genotypes)
}
\keyword{datasets}
